+++ oe_patterns_cdk_common/db_secret.py
@@ -17,6 +17,7 @@ class DbSecret(Construct):
             self,
             scope: Construct,
             id: str,
+            username: str = 'dbadmin',
             **props):
         super().__init__(scope, id, **props)
 
@@ -45,10 +46,10 @@ class DbSecret(Construct):
             self,
             "DbSecret",
             generate_secret_string=aws_secretsmanager.CfnSecret.GenerateSecretStringProperty(
-                exclude_characters="\"@/\\\"'$,[]*?{}~\#%<>|^",
+                exclude_characters="\"@/\\\"'$,[]*?{}~#%<>|^",
                 exclude_punctuation=True,
                 generate_string_key="password",
-                secret_string_template=json.dumps({"username":"dbadmin"})
+                secret_string_template=json.dumps({"username":username})
             ),
             name="{}/db/secret".format(Aws.STACK_NAME)
         )