+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Unreleased
 
+# 3.8.0
+
+* fixing extra test output
+* fixing ses secretvalue permissions issue
+* allowing dbadmin username to be customized in DbSecret construct
+* switch from launch configs to launch templates in Asg construct
+* allow customization of asg health check type
+* allowing default database to be customized in AuroraCluster construct
+* allow customization of size of root volume
+* [Apex DNS issue](https://github.com/ordinaryexperts/aws-marketplace-oe-patterns-cdk-common/issues/5)
+
 # 3.7.0
 
 * fixing AssetsBucket when a bucket name is provided